Code clean-up and improved documentation

+++ src/Command/DatabaseDropCommand.php
@@ -2,15 +2,14 @@
 
 namespace App\Command;
 
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Doctrine\DBAL\Connection;
 
 /**
- * Creates a database
+ * Commands are configured as service in app/config/console.yml
  */
-class DatabaseDropCommand extends Command
+class DatabaseDropCommand extends AbstractCommand
 {
     protected $connection;
 
@@ -19,13 +18,29 @@ class DatabaseDropCommand extends Command
         $this->connection = $connection;
 
         parent::__construct($name);
+
+        $this->setDescription('Drops the database configured in app/config/parameters.yml');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $db = $this->connection;
 
-        $db->exec("DROP DATABASE `{$db->getDatabase()}`");
+        $question = 'Are you sure you wish to drop the database "' . $db->getDatabase() . '"? (y/n)';
+
+        if (! $this->canExecute($question, $input, $output)) {
+            $output->writeln('<error>Database "' . $db->getDatabase() . '" was not dropped</error>');
+
+            return 1;
+        }
+
+        try {
+            $db->exec("DROP DATABASE `{$db->getDatabase()}`");
+        } catch (\Exception $e) {
+            $output->writeln('<error>' . $e->getMessage() . '</error>');
+            return 1;
+        }
+
         $output->writeln('<info>Database "' . $db->getDatabase() . '" dropped</info>');
     }
 }
